+++ src/systems.js
@@ -19,6 +19,7 @@ import knave from "./systems/knave.js";
 import t2k4e from "./systems/t2k4e.js";
 import yzecoriolis from "./systems/yzecoriolis.js";
 import kamigakari from "./systems/kamigakari.js";
+import ose from "./systems/ose.js";
 // ↑ IMPORT SYSTEMS HERE ↑
 
 /**
@@ -49,6 +50,7 @@ export const SYSTEMS = {
     "t2k4e": t2k4e,
     "yzecoriolis": yzecoriolis,
     "kamigakari": kamigakari,
+    //"ose": ose,
     // ↑ ADD SYSTEMS HERE ↑
   },